Move Session and Config to xlib, split off PriorityList
* Move Input.AddUser + lock/uids/uidComplete/lastTarget to Session where it belongs
* Add functions to Session to give access to these
* Split mainloop into two parts (Session and UI)
* Many many small changes to get things compiling again ;)

but it still works after all this :)
